Ensure OAuth is re-seeded by db-migrations container
