+++ backend/LexBoxApi/Auth/AuthKernel.cs
@@ -33,7 +33,7 @@ public static class AuthKernel
 
     public static void AddLexBoxAuth(IServiceCollection services,
         IConfigurationRoot configuration,
-        IWebHostEnvironment environment)
+        IHostEnvironment environment)
     {
         if (environment.IsDevelopment())
         {
@@ -241,7 +241,7 @@ public static class AuthKernel
         services.AddOptions<AuthenticationOptions>().ValidateOnStart();
     }
 
-    private static void AddOpenId(IServiceCollection services, IWebHostEnvironment environment)
+    public static void AddOpenId(IServiceCollection services, IHostEnvironment environment)
     {
         services.Add(ScopeRequestFixer.Descriptor.ServiceDescriptor);
         //openid server
